support GPX 1.1 and validate against schema

+++ lib/src/main/src/beeline/co/gpx/Elevation.kt
@@ -4,7 +4,7 @@ import beeline.co.gpx.xml.XmlWritable
 import beeline.co.gpx.xml.XmlWrite
 import rx.Observable
 
-data class Elevation(val elevation: Int?) : XmlWritable {
+data class Elevation(val elevation: Double?) : XmlWritable {
 
     override val writeOperations: Observable<XmlWrite>
         get() = if (elevation != null) newTag("ele", withText(elevation.toString()))
